🏗️ Architect: Modularized PlaceConditionsMixIn
Decomposed the legacy `apts/place/conditions.py` God Mixin into a modular package structure in `apts/place/conditions/`.

Key changes:
- Created specialized mixins for Light Pollution, SQM estimation, Seeing estimation, and Weather integration.
- Extracted vectorized bulk calculations into internal methods within their respective mixins, improving readability of `_add_extra_weather_info`.
- Recomposed `PlaceConditionsMixIn` using multiple inheritance in `apts/place/conditions/base.py`.
- Maintained backward compatibility via re-exports in `__init__.py`.
- Updated all dependent modules (`times.py`, `imaging.py`, `paths.py`) to the new import path.
- Verified integrity with 80 unit tests and linting.
